backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# Cache
redis[hiredis]==5.2.1

# Config & validation
pydantic==2.10.4
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib==1.7.4

# HTTP client
httpx==0.28.1

# Email
aiosmtplib==3.0.2
jinja2==3.1.5

# Logging
structlog==24.4.0

# Rate limiting
slowapi==0.1.9

# LLM clients
# anthropic >= 0.49 is required for the `thinking={"type": "enabled", ...}`
# kwarg used by llm_service. 0.42 silently rejects it as an unexpected
# keyword and the call falls through to the OpenAI fallback. See
# LEARNINGS L4.9.
anthropic==0.49.0
openai==1.58.1

# Cross-encoder reranking
sentence-transformers==3.3.1

# Circuit breaker (LLM fallback)
pybreaker==1.2.0

# File uploads (required by FastAPI UploadFile)
python-multipart==0.0.18

# Utilities
tenacity==9.0.0
tiktoken==0.8.0
email-validator==2.2.0
aiodns==3.2.0

# PDF export + QR codes (compliance brief download)
reportlab==4.2.5
qrcode[pil]==7.4.2
